+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changes in version 0.9 (dev)
 * Allow ``formfield_overrides`` to contain field names too.
 * API: added ``SeoPageMixin`` model with ``meta_title``, ``meta_keywords`` and ``meta_description`` fields.
 * API: renamed ``FluentPageBase`` to ``AbstractFluentPage``.
+* API: added ``get_view_response`` to the ``PageTypePlugin`` class, allow adding middleware to custom views.
 * API: **Backwards incompatible:** when inheriting from the abstract ``HtmlPage`` model, your app needs a South migration.
 * Dropped Django 1.3 support.